+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Connection::exist_by_pk`
+- `Connection::exist_one`
 
 ## [5.0.0] - 2025-10-26